fix bug by change double to int

+++ src/main/java/Controller/CustomerController.java
@@ -183,7 +183,7 @@ public class CustomerController {
                 .collect(Collectors.toList());
     }
 
-    public void rateProduct(User user, String productId, Double rate) throws Exception {
+    public void rateProduct(User user, String productId, int rate) throws Exception {
         Product product = ProductController.getProductById(productId);
         if (!((Customer) user).getRecentShoppingProducts().contains(product))
             throw new Exception("Customer Does'nt buy this Product");
